+++ src/radio/sx1276/sx1276.h
@@ -395,7 +395,7 @@ int16_t SX1276ReadRssi( RadioModems_t modem );
  * \param [IN]: addr Register address
  * \param [IN]: data New register value
  */
-void SX1276Write( uint16_t addr, uint8_t data );
+void SX1276Write( uint32_t addr, uint8_t data );
 
 /*!
  * \brief Reads the radio register at the specified address
@@ -403,7 +403,7 @@ void SX1276Write( uint16_t addr, uint8_t data );
  * \param [IN]: addr Register address
  * \retval data Register value
  */
-uint8_t SX1276Read( uint16_t addr );
+uint8_t SX1276Read( uint32_t addr );
 
 /*!
  * \brief Writes multiple radio registers starting at address
@@ -412,7 +412,7 @@ uint8_t SX1276Read( uint16_t addr );
  * \param [IN] buffer Buffer containing the new register's values
  * \param [IN] size   Number of registers to be written
  */
-void SX1276WriteBuffer( uint16_t addr, uint8_t *buffer, uint8_t size );
+void SX1276WriteBuffer( uint32_t addr, uint8_t *buffer, uint8_t size );
 
 /*!
  * \brief Reads multiple radio registers starting at address
@@ -421,7 +421,7 @@ void SX1276WriteBuffer( uint16_t addr, uint8_t *buffer, uint8_t size );
  * \param [OUT] buffer Buffer where to copy the registers data
  * \param [IN] size Number of registers to be read
  */
-void SX1276ReadBuffer( uint16_t addr, uint8_t *buffer, uint8_t size );
+void SX1276ReadBuffer( uint32_t addr, uint8_t *buffer, uint8_t size );
 
 /*!
  * \brief Sets the maximum payload length.
